+++ src/main/java/jen/JenException.java
@@ -1,3 +1,4 @@
+package jen;
 public class JenException extends Exception{
     public JenException(String msg) {
         super(msg);
