+++ src/main/java/jen/OutOfIndexException.java
@@ -1,3 +1,4 @@
+package jen;
 public class OutOfIndexException extends Exception{
     public OutOfIndexException(String msg) {
         super(msg);
