+++ src/main/java/jen/UI.java
@@ -1,3 +1,4 @@
+package jen;
 import java.util.Scanner;
 public class UI {
     private Scanner scanner;
