+++ src/main/java/jen/commands/ByeCommand.java
@@ -1,3 +1,7 @@
+package jen.commands;
+import jen.UI;
+import jen.Storage;
+
 public class ByeCommand extends Command {
     @Override
     public void run(Storage store, UI ui) {
