+++ src/main/java/jen/commands/Command.java
@@ -1,3 +1,8 @@
+package jen.commands;
+import jen.OutOfIndexException;
+import jen.Storage;
+import jen.UI;
+
 public abstract class Command {
     public abstract void run(Storage store, UI ui) throws OutOfIndexException;
 }
