+++ src/main/java/jen/commands/CommandType.java
@@ -1,3 +1,4 @@
+package jen.commands;
 public enum CommandType {
     TODO, DEADLINE, EVENT, LIST, MARK, UNMARK, DELETE, BYE, UNKNOWN
 }
