+++ src/main/java/jen/commands/DeleteCommand.java
@@ -1,8 +1,11 @@
-public class MarkCommand extends Command {
+package jen.commands;
+import jen.OutOfIndexException;
+import jen.Storage;
+import jen.UI;
 
+public class DeleteCommand extends Command {
     private int index;
-    public MarkCommand(int i) {
-
+    public DeleteCommand(int i) {
         this.index = i;
     }
     @Override
@@ -10,8 +13,7 @@ public class MarkCommand extends Command {
         if (!storage.isWithinSize(this.index)) {
             throw new OutOfIndexException("Input index outside of list size");
         }
-        storage.markAsDone(index);
-        ui.printMessage("I have marked task as done!\n" + storage.taskToString(index) +
+        ui.printMessage("I have removed this task:\n" + storage.deleteItem(this.index) +
                 "\n" + storage.sizeToString());
     }
 }
