+++ src/main/java/jen/commands/MarkCommand.java
@@ -1,6 +1,14 @@
-public class DeleteCommand extends Command {
+package jen.commands;
+
+import jen.OutOfIndexException;
+import jen.Storage;
+import jen.UI;
+
+public class MarkCommand extends Command {
+
     private int index;
-    public DeleteCommand(int i) {
+    public MarkCommand(int i) {
+
         this.index = i;
     }
     @Override
@@ -8,7 +16,8 @@ public class DeleteCommand extends Command {
         if (!storage.isWithinSize(this.index)) {
             throw new OutOfIndexException("Input index outside of list size");
         }
-        ui.printMessage("I have removed this task:\n" + storage.deleteItem(this.index) +
+        storage.markAsDone(index);
+        ui.printMessage("I have marked task as done!\n" + storage.taskToString(index) +
                 "\n" + storage.sizeToString());
     }
 }
