+++ src/main/java/jen/tasks/Event.java
@@ -1,3 +1,4 @@
+package jen.tasks;
 public class Event extends Task{
     protected String from;
     protected String to;
